parse subscribe

+++ akamq-mqtt/src/main/java/com/github/sioncheng/akamq/mqtt/MQTTMessageType.java
@@ -6,35 +6,35 @@ package com.github.sioncheng.akamq.mqtt;
  */
 public class MQTTMessageType {
 
-    public static final int RESERVED = 0;
+    public static final byte RESERVED = 0;
 
-    public static final int CONNECT = 1;
+    public static final byte CONNECT = 1;
 
-    public static final int CONNECT_ACK = 2;
+    public static final byte CONNECT_ACK = 2;
 
-    public static final int PUBLISH  = 3;
+    public static final byte PUBLISH  = 3;
 
-    public static final int PUBLISH_ACK = 4;
+    public static final byte PUBLISH_ACK = 4;
 
-    public static final int PUBLISH_RECEIVED = 5;
+    public static final byte PUBLISH_RECEIVED = 5;
 
-    public static final int PUBLISH_RELEASE = 6;
+    public static final byte PUBLISH_RELEASE = 6;
 
-    public static final int PUBLISH_COMPLETE = 7;
+    public static final byte PUBLISH_COMPLETE = 7;
 
-    public static final int SUBSCRIBE = 8;
+    public static final byte SUBSCRIBE = 8;
 
-    public static final int SUBSCRIBE_ACK = 9;
+    public static final byte SUBSCRIBE_ACK = 9;
 
-    public static final int UN_SUBSCRIBE = 10;
+    public static final byte UN_SUBSCRIBE = 10;
 
-    public static final int UN_SUBSCRIBE_ACK = 11;
+    public static final byte UN_SUBSCRIBE_ACK = 11;
 
-    public static final int PING_REQUEST = 12;
+    public static final byte PING_REQUEST = 12;
 
-    public static final int PING_RESPONSE = 13;
+    public static final byte PING_RESPONSE = 13;
 
-    public static final int DISCONNECT = 14;
+    public static final byte DISCONNECT = 14;
 
-    public static final int RESERVED_15 = 15;
+    public static final byte RESERVED_15 = 15;
 }
